+++ feature/diary/src/main/java/com/boostcamp/dreamteam/dreamdiary/feature/diary/home/DiaryHomeUIState.kt
@@ -1,0 +1,8 @@
+package com.boostcamp.dreamteam.dreamdiary.feature.diary.home
+
+import com.boostcamp.dreamteam.dreamdiary.core.model.Diary
+
+data class DiaryHomeUIState(
+    val diaries: List<Diary> = listOf(),
+    val loading: Boolean = false,
+)
